+++ clippy_utils/src/str_utils.rs
